backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv==1.0.0
email-validator==2.1.0
PyMySQL==1.1.0
asyncpg==0.30.0
